ModelManagerStudio

Documentation for ModelManagerStudio.

ModelManagerStudio.get_pcmm_paths

ModelManagerStudio.init_model_manager_gui

ModelManagerStudio.launch

ModelManagerStudio.main

ModelManagerStudio.studio_initialize_model_manager

ModelManagerStudio.get_pcmm_paths — Method

get_pcmm_paths(args::Vararg{AbstractString})

Get the PhysiCell and data directories from the provided arguments. See launch for details on how the arguments are interpreted.

ModelManagerStudio.init_model_manager_gui — Method

init_model_manager_gui(args::Vararg{AbstractString}; kwargs...)

Internal function to initialize the Model Manager GUI with the specified arguments. Called by launch.

ModelManagerStudio.launch — Method

launch()
launch(path_to_project::AbstractString)
launch(physicell_dir::AbstractString, data_dir::AbstractString)

Launch the Model Manager Studio GUI for a particular project.

If no arguments are provided, it will initialize the Model Manager within the current working directory, i.e., using data_dir = "./data" and physicell_dir = "./PhysiCell". If a single argument is provided, it will use that as the base directory and append data and PhysiCell to it, i.e., data_dir = "$(path_to_project)/data" and physicell_dir = "$(path_to_project)/PhysiCell". If two arguments are provided, they will be used as the physicell_dir and data_dir, respectively.

Examples

using ModelManagerStudio
launch()  # Launches with default paths in the current directory, i.e., `data_dir = "./data"` and `physicell_dir = "./PhysiCell"`
launch("path/to/project")  # Launches with specified project path (will ask if paths differ)
launch("path/to/PhysiCell", "path/to/data")  # Launches with specified PhysiCell and data directories

ModelManagerStudio.main — Method

main(args)

Entry point for ModelManagerStudio application. Calls launch to start the application. See launch for argument details.

ModelManagerStudio.studio_initialize_model_manager — Method

studio_initialize_model_manager(args::Vararg{AbstractString})

Internal function to initialize the Model Manager with the specified arguments. Called by init_model_manager_gui.